separate layouts from widgets
