+++ src/widgets/src/index.ts
@@ -1,0 +1,27 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) 2014-2016, PhosphorJS Contributors
+|
+| Distributed under the terms of the BSD 3-Clause License.
+|
+| The full license is in the file LICENSE, distributed with this software.
+|----------------------------------------------------------------------------*/
+export * from './boxengine';
+export * from './boxlayout';
+export * from './boxpanel';
+export * from './commandpalette';
+//export * from './dockpanel';
+export * from './domutil';
+//export * from './focustracker';
+export * from './layout';
+export * from './menu';
+export * from './menubar';
+export * from './panel';
+export * from './panellayout';
+export * from './splitlayout';
+export * from './splitpanel';
+export * from './stackedlayout';
+export * from './stackedpanel';
+export * from './tabbar';
+export * from './tabpanel';
+export * from './title';
+export * from './widget';
